Small changes regarding android
1, added flags in cmake file to supress warning while compiling.
2. added casting func
3. changed color to make visible setting option in android app

+++ code/mobile/android/PhoneVR/app/src/main/cpp/native-lib.cpp
@@ -165,7 +165,7 @@ SUB(startMediaCodec)(JNIEnv *env, jclass, jobject surface) {
                 {
                     size_t bufSz = 0;
                     uint8_t *buf = AMediaCodec_getInputBuffer(codec, (size_t)idx, &bufSz);
-                    PVREnqueueVideoBuf({buf, idx, bufSz});
+                    PVREnqueueVideoBuf({buf, static_cast<int>(idx), bufSz});
                 }
             }
